+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed 
 - History does not contain duplicate items
 
+### Fixed
+- Invalid history records with pod name `<none>`
+
 ## [1.2.0] - 2020-06-16
 ### Added
 - Option for connecting to Cloud SQL instance with instance name only